refactor groupby_and_aggregate
