+++ bodo/ir/aggregate.py
@@ -49,7 +49,6 @@ from bodo.libs.array import (
     cpp_table_to_py_data,
     delete_info,
     delete_table,
-    groupby_and_aggregate,
     py_data_to_cpp_table,
 )
 from bodo.libs.array_item_arr_ext import (
@@ -1458,6 +1457,132 @@ class EvalDummyTyper(AbstractTemplate):
         assert not kws
         # takes the output array as first argument to know the output dtype
         return signature(args[0].dtype, *args)
+
+
+@intrinsic
+def groupby_and_aggregate(
+    typingctx,
+    table_t,
+    n_keys_t,
+    cols_per_func_t,
+    nwindows_calls_per_func_t,
+    n_funcs_t,
+    input_has_index,
+    ftypes,
+    func_offsets,
+    udf_n_redvars,
+    is_parallel,
+    skip_na_data_t,
+    shift_periods_t,
+    transform_func,
+    head_n,
+    return_keys,
+    return_index,
+    dropna,
+    update_cb,
+    combine_cb,
+    eval_cb,
+    general_udfs_cb,
+    udf_table_dummy_t,
+    n_out_rows_t,
+    window_ascending_t,
+    window_na_position_t,
+    window_args_t,
+    n_window_args_per_func_t,
+    n_input_cols_per_func_t,
+    maintain_input_size_t,
+    n_shuffle_keys_t,
+    use_sql_rules_t,
+):
+    """
+    Interface to groupby_and_aggregate function in C++ library for groupby
+    offloading.
+    """
+    from bodo.libs.array import table_type
+
+    assert table_t == table_type
+    assert udf_table_dummy_t == table_type
+
+    def codegen(context, builder, sig, args):
+        fnty = lir.FunctionType(
+            lir.IntType(8).as_pointer(),  # table_info*
+            [
+                lir.IntType(8).as_pointer(),
+                lir.IntType(64),
+                lir.IntType(8).as_pointer(),
+                lir.IntType(8).as_pointer(),
+                lir.IntType(64),
+                lir.IntType(1),
+                lir.IntType(8).as_pointer(),
+                lir.IntType(8).as_pointer(),
+                lir.IntType(8).as_pointer(),
+                lir.IntType(1),
+                lir.IntType(1),
+                lir.IntType(64),  # shift_periods_t
+                lir.IntType(8).as_pointer(),  # transform_func
+                lir.IntType(64),  # head_n
+                lir.IntType(1),
+                lir.IntType(1),
+                lir.IntType(1),  # groupby key dropna
+                lir.IntType(8).as_pointer(),
+                lir.IntType(8).as_pointer(),
+                lir.IntType(8).as_pointer(),
+                lir.IntType(8).as_pointer(),
+                lir.IntType(8).as_pointer(),
+                lir.IntType(8).as_pointer(),
+                lir.IntType(8).as_pointer(),  # window_ascending
+                lir.IntType(8).as_pointer(),  # window_na_position
+                lir.IntType(8).as_pointer(),  # window_args
+                lir.IntType(8).as_pointer(),  # n_window_args_per_func
+                lir.IntType(8).as_pointer(),  # n_input_cols_per_func
+                lir.IntType(1),  # maintain_input_size
+                lir.IntType(64),  # n_shuffle_keys_t
+                lir.IntType(1),  # use_sql_rules_t
+            ],
+        )
+        fn_tp = cgutils.get_or_insert_function(
+            builder.module, fnty, name="groupby_and_aggregate"
+        )
+        ret = builder.call(fn_tp, args)
+        bodo.utils.utils.inlined_check_and_propagate_cpp_exception(context, builder)
+        return ret
+
+    return (
+        table_type(
+            table_t,
+            types.int64,
+            types.voidptr,
+            types.voidptr,
+            types.int64,
+            types.boolean,
+            types.voidptr,
+            types.voidptr,
+            types.voidptr,
+            types.boolean,
+            types.boolean,
+            types.int64,  # shift_periods
+            types.voidptr,  # transform_func
+            types.int64,  # head_n
+            types.boolean,
+            types.boolean,
+            types.boolean,  # dropna
+            types.voidptr,
+            types.voidptr,
+            types.voidptr,
+            types.voidptr,
+            table_t,
+            types.voidptr,
+            types.voidptr,  # window_ascending
+            types.voidptr,  # window_na_position
+            window_args_t,  # window_args
+            types.voidptr,  # n_window_args_per_func
+            types.voidptr,  # n_input_cols_per_func
+            types.boolean,  # maintain_input_size
+            types.int64,  # n_shuffle_keys_t
+            types.boolean,  # use_sql_rules_t
+        ),
+        codegen,
+    )
 
 
 def agg_distributed_run(
